Now compiles and run if run from same folder as original LM-file
